Move method to student.service.ts
