+++ src/web/services/student.service.spec.ts
@@ -84,4 +84,14 @@ describe('StudentService', () => {
     expect(spyHttpRequestService.get)
         .toHaveBeenCalledWith(ResourceEndpoints.STUDENTS_CSV, paramMap, responseType);
   });
+
+  it('should execute GET when getting students enrolled in course', () => {
+    const paramMap: Record<string, string> = {
+      courseid: 'CS3281',
+    };
+    service.getStudentsEnrolledInCourse({
+      courseId: paramMap.courseid,
+    });
+    expect(spyHttpRequestService.get).toHaveBeenCalledWith(ResourceEndpoints.STUDENTS, paramMap);
+  });
 });
